refactor(config): split config_test.go into focused test files
Splits the 552-line config_test.go into four focused files:

- config_arch_test.go — NormalizeArch, ServerPlatformDir, BinariesPlatformDir, UEPlatformName
- config_defaults_test.go — TestDefaults (all string, numeric, map checks as named subtests)
- config_game_test.go — GameConfig method tests (resolved targets, arch, project path)
- config_load_test.go — Load/YAML tests including DDC config and deprecated lyra key

config_test.go is reduced to the single cross-cutting TestResolveServerBuildDir.

Quality fixes: TestLoad_DeprecatedLyraKey migrated to t.Chdir(), numeric/map
assertions in TestDefaults promoted to named t.Run subtests, TestLoad_ValidYAML
field checks consolidated into a table.
